+++ functions/Export-MyTimer.ps1
@@ -1,5 +1,5 @@
 Function Export-MyTimer {
-    [cmdletbinding(SupportsShouldProcess)]
+    [CmdletBinding(SupportsShouldProcess)]
     [OutputType('None')]
     Param(
         [Parameter(