+++ functions/Restart-MyTimer.ps1
@@ -1,5 +1,5 @@
 Function Restart-MyTimer {
-    [cmdletbinding(SupportsShouldProcess)]
+    [CmdletBinding(SupportsShouldProcess)]
     [OutputType('None','MyTimer')]
     Param(
         [Parameter(